requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
environs==14.1.1
marshmallow==3.26.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
